Move Throttler to breadboard.
